put serverside get and post methods to separate files
